requirements.txt
pbr
six
future
cython>=0.24.0a0
numpy>=1.10.4
scipy>=0.13.3
flask
pycrypto
pyenchant